+++ src/main/java/org/thisway/company/controller/dto/request/AdminCompanyUpdateRequest.java
@@ -1,11 +1,11 @@
-package org.thisway.company.dto.request;
+package org.thisway.company.controller.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Positive;
-import org.thisway.company.dto.AdminCompanyRegisterInput;
+import org.thisway.company.service.dto.input.AdminCompanyUpdateInput;
 
-public record AdminCompanyRegisterRequest(
+public record AdminCompanyUpdateRequest(
+
         @NotBlank
         String name,
 
@@ -22,21 +22,18 @@ public record AdminCompanyRegisterRequest(
         String addrDetail,
 
         @NotNull
-        String memo,
-
-        @Positive
-        Integer gpsCycle
+        String memo
 ) {
 
-    public AdminCompanyRegisterInput toCompanyRegisterInput() {
-        return AdminCompanyRegisterInput.builder()
+    public AdminCompanyUpdateInput toCompanyUpdateInput(long id) {
+        return AdminCompanyUpdateInput.builder()
+                .id(id)
                 .name(name)
                 .crn(crn)
                 .contact(contact)
                 .addrRoad(addrRoad)
                 .addrDetail(addrDetail)
                 .memo(memo)
-                .gpsCycle(gpsCycle)
                 .build();
     }
 }
