+++ src/main/java/org/thisway/company/service/dto/input/AdminCompanyRegisterInput.java
@@ -1,4 +1,4 @@
-package org.thisway.company.dto;
+package org.thisway.company.service.dto.input;
 
 import lombok.Builder;
 import org.thisway.company.entity.Company;
