+++ src/main/java/org/thisway/company/service/dto/output/AdminCompaniesOutput.java
@@ -1,4 +1,4 @@
-package org.thisway.company.dto;
+package org.thisway.company.service.dto.output;
 
 
 import java.util.List;
