+++ src/test/java/org/thisway/company/service/AdminCompanyServiceTest.java
@@ -18,9 +18,9 @@ import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 import org.thisway.common.CustomException;
 import org.thisway.common.ErrorCode;
-import org.thisway.company.dto.AdminCompaniesOutput;
-import org.thisway.company.dto.AdminCompanyDetailOutput;
-import org.thisway.company.dto.AdminCompanyRegisterInput;
+import org.thisway.company.service.dto.output.AdminCompaniesOutput;
+import org.thisway.company.service.dto.output.AdminCompanyDetailOutput;
+import org.thisway.company.service.dto.input.AdminCompanyRegisterInput;
 import org.thisway.company.entity.Company;
 import org.thisway.company.repository.CompanyRepository;
 import org.thisway.company.support.CompanyFixture;
